Add files via upload
Adding my own ideas to this shared folder (to avoid too many folder with just ideas). Some are a bit too ambitious or not necessarily useful to visualise, but it was the first outcast. There is a hierarchy to the list - ideas are sorted from most interesting/stimulating/possible to less possible/too ambitious.